requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
pandas==2.2.3
azure-storage-blob==12.26.0
azure-search-documents==11.6.0
azure-cosmos==4.9.0
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
openai==1.55.3
httpx==0.28.1
python-multipart==0.0.20
rapidfuzz==3.12.2
metaphone==0.6
slowapi==0.1.9
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
bcrypt==4.2.1
itsdangerous==2.2.0
